requirements.txt
faiss-gpu
jax
jaxlib
jax-cuda
loguru
umap-learn
pandas
numpy
optuna
scikit-learn
scikit-learn-extra
scikit-learn-contrib
openml
scipy
scikit-optimize
